requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
chromadb
pdfplumber
python-dotenv
tqdm
sentence-transformers
google-api-python-client
google-auth
google-auth-oauthlib

